improve unit tests for aggregate rewriting
